requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
resend>=2.0.0
